Balance Patch
총알 데미지 계산을 합연산에서 곱연산으로 변경
일부 총 데미지 너프
업그레이드 UI로 스탯이 정상적으로 업그레이드 되게 변경

+++ RogueLike_Project/Assets/2. Scripts/UIScript/UIDamage.cs
@@ -15,7 +15,7 @@ public class UIDamage : MonoBehaviour
     void Start()
     {
         //text = GetComponentInChildren<TMP_Text>();
-        text.text = damage.ToString();
+        text.text = ((int)damage).ToString();
         Destroy(gameObject, 1f);
         player = FindObjectOfType<PlayerControl>().gameObject.transform;
     }
